feat(ws): change invalid session handling
Change behavior of the Session upon receiving Invalid Session opcode.
Previously in such case, the Session would re-identify on the existing connection.
However due to introduction of separate resume gateway and updated connection
lifecycle, session now needs to reconnect to the gateway.
In case connection is not resumable, gateway session details will be
dropped and the Session would re-identify, but on a new connection.